allow config location to be specified with an encironment variable
