+++ .github/workflows/integration.yml
@@ -97,7 +97,16 @@ jobs:
             --pattern "*network*.py" \
             --cfn-stack-name ${{ needs.generate-stack-name.outputs.stack-name }} \
             --aws-region ${{ secrets.AWS_REGION }}
+      - name: Run Notebook Tests
+        run: |
+          python test/integration/NeptuneIntegrationWorkflowSteps.py \
+            run-tests \
+            --pattern "*graph_notebook.py" \
+            --cfn-stack-name ${{ needs.generate-stack-name.outputs.stack-name }} \
+            --aws-region ${{ secrets.AWS_REGION }}
       - name: Run IAM Tests
+        env:
+          GRAPH_NOTEBOK_CONFIG: /tmp/graph_notebook_config.json
         run: |
           python test/integration/NeptuneIntegrationWorkflowSteps.py \
             run-tests \
