--- test/integration/graph_notebook/test_graph_notebook_gremlin.py
@@ -1,4 +1,0 @@
-"""
-Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-SPDX-License-Identifier: Apache-2.0
-"""
