+++ test/integration/graph_notebook/test_gremlin_graph_notebook.py
@@ -6,14 +6,10 @@ SPDX-License-Identifier: Apache-2.0
 from test.integration.graph_notebook.GraphNotebookIntegrationTest import GraphNotebookIntegrationTest
 
 
-class TestGraphMagic(GraphNotebookIntegrationTest):
+class TestGraphMagicGremlin(GraphNotebookIntegrationTest):
     def tearDown(self) -> None:
         delete_query = "g.V('graph-notebook-test').drop()"
         self.ip.run_cell_magic('gremlin', 'query', delete_query)
-
-    def test_status(self):
-        res = self.ip.run_line_magic('status', '')
-        self.assertEqual('healthy', res['status'])
 
     def test_gremlin_query(self):
         label = 'graph-notebook-test'
@@ -21,6 +17,7 @@ class TestGraphMagic(GraphNotebookIntegrationTest):
 
         store_to_var = 'gremlin_res'
         self.ip.run_cell_magic('gremlin', f'query --store-to {store_to_var}', query)
+        self.assertFalse('graph_notebook_error' in self.ip.user_ns)
         gremlin_res = self.ip.user_ns[store_to_var]
 
         # TODO: how can we get a look at the objects which were displayed?
